Foi adicionado o Arquivo Conftest ao projeto.
close #61
