+++ libpythonpro/test/test_spam/test_enviador_de_email.py
@@ -1,5 +1,4 @@
 import pytest
-
 from libpythonpro.spam.enviador_de_email import Enviador, EmailEnvalido
